requirements.txt
aioredis==1.3.1
arrow==1.2.3
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
backports.zoneinfo==0.2.1
binaryornot==0.4.4
certifi==2022.9.24
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
constantly==15.1.0
cookiecutter==2.1.1
cryptography==37.0.4
daphne==3.0.2
dj-database-url==2.0.0
Django==4.1
django-browser-reload==1.6.0
django-phonenumber-field==6.4.0
django-phonenumbers==1.0.1
django-tailwind==3.5.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.1
msgpack==1.0.4
phonenumbers==8.12.54
Pillow==9.2.0
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
python-dateutil==2.8.2
python-decouple==3.6
python-slugify==6.1.2
PyYAML==6.0
requests==2.28.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
urllib3==1.26.12
zope.interface==5.4.0
